+++ build.gradle
@@ -48,11 +48,11 @@ test {
 }
 
 application {
-    mainClass.set("duke.Launcher")
+    mainClass.set("alice.Launcher")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "alice"
     archiveClassifier = null
     dependsOn("distZip", "distTar")
 }
