+++ src/main/java/alice/Launcher.java
@@ -1,0 +1,17 @@
+package alice;
+
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    /**
+     * Launches the application.
+     *
+     * @param args The arguments passed in.
+     */
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
