+++ src/main/java/alice/Main.java
@@ -1,0 +1,36 @@
+package alice;
+
+import java.io.IOException;
+
+import alice.ui.components.MainWindow;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+
+/**
+ * A GUI for Alice using FXML.
+ */
+public class Main extends Application {
+    // The instance of Alice that will be used throughout the program
+    private final Alice alice = new Alice();
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            fxmlLoader.<MainWindow>getController().setDuke(alice);
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
+
