+++ src/main/java/alice/command/DeleteCommand.java
@@ -1,0 +1,34 @@
+package alice.command;
+
+import alice.exception.DukeException;
+import alice.storage.Storage;
+import alice.task.Task;
+import alice.task.TaskList;
+import alice.ui.Ui;
+
+/**
+ * Represents a command issued by the user to delete a task from the list.
+ */
+public class DeleteCommand extends Command {
+    private final int index; // The index of the task to be deleted.
+
+    /**
+     * Constructs a DeleteCommand with the given index.
+     *
+     * @param index The index of the task to be deleted.
+     */
+    public DeleteCommand(int index) {
+        this.index = index;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        Task task = tasks.get(this.index);
+        tasks.delete(this.index);
+        storage.save(tasks.toFileString());
+        return ui.showDeleteTask(task, tasks.size());
+    }
+}
