+++ src/main/java/alice/command/MarkAsDoneCommand.java
@@ -1,0 +1,32 @@
+package alice.command;
+
+import alice.exception.DukeException;
+import alice.storage.Storage;
+import alice.task.TaskList;
+import alice.ui.Ui;
+
+/**
+ * Represents a command issued by the user to mark a task as done.
+ */
+public class MarkAsDoneCommand extends Command {
+    private final int index; // The index of the task to be marked as done.
+
+    /**
+     * Constructs a MarkAsDoneCommand with the given index.
+     *
+     * @param index The index of the task to be marked as done.
+     */
+    public MarkAsDoneCommand(int index) {
+        this.index = index;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        tasks.markAsDone(this.index);
+        storage.save(tasks.toFileString());
+        return ui.showMarkTaskAsDone(tasks.get(this.index));
+    }
+}
