+++ src/main/java/alice/command/UnmarkAsDoneCommand.java
@@ -1,0 +1,32 @@
+package alice.command;
+
+import alice.exception.DukeException;
+import alice.storage.Storage;
+import alice.task.TaskList;
+import alice.ui.Ui;
+
+/**
+ * Represents a command issued by the user to mark a task as not done.
+ */
+public class UnmarkAsDoneCommand extends Command {
+    private final int index; // The index of the task to be marked as not done.
+
+    /**
+     * Constructs a UnmarkAsDoneCommand with the given index.
+     *
+     * @param index The index of the task to be marked as not done.
+     */
+    public UnmarkAsDoneCommand(int index) {
+        this.index = index;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        tasks.unmarkAsDone(this.index);
+        storage.save(tasks.toFileString());
+        return ui.showUnmarkTaskAsDone(tasks.get(this.index));
+    }
+}
