--- src/main/java/duke/command/ExitCommand.java
@@ -1,26 +1,0 @@
-package duke.command;
-
-import duke.exception.DukeException;
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command issued by the user to close the application.
- */
-public class ExitCommand extends Command {
-    /**
-     * Constructs an ExitCommand.
-     */
-    public ExitCommand() {
-        this.isExit = true;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        return ui.showGoodbye();
-    }
-}
