--- src/main/java/duke/command/ListCommand.java
@@ -1,19 +1,0 @@
-package duke.command;
-
-import duke.exception.DukeException;
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command issued by the user to list all tasks in the list.
- */
-public class ListCommand extends Command {
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        return ui.showTaskList(tasks);
-    }
-}
