--- src/main/java/duke/command/MarkAsDoneCommand.java
@@ -1,32 +1,0 @@
-package duke.command;
-
-import duke.exception.DukeException;
-import duke.storage.Storage;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command issued by the user to mark a task as done.
- */
-public class MarkAsDoneCommand extends Command {
-    private final int index; // The index of the task to be marked as done.
-
-    /**
-     * Constructs a MarkAsDoneCommand with the given index.
-     *
-     * @param index The index of the task to be marked as done.
-     */
-    public MarkAsDoneCommand(int index) {
-        this.index = index;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        tasks.markAsDone(this.index);
-        storage.save(tasks.toFileString());
-        return ui.showMarkTaskAsDone(tasks.get(this.index));
-    }
-}
